+++ src/components/loader/loading-styles.js
@@ -1,0 +1,11 @@
+import { makeStyles } from "@material-ui/core";
+
+export const useStyles = makeStyles(() => ({
+  container: {
+    minWidth: "100vw",
+    minHeight: "100vh",
+    display: "flex",
+    justifyContent: "center",
+    alignItems: "center",
+  },
+}));
